refactor: split TelegramBot into modular components
Split TelegramChannel.js (1573 LOC) into focused modules:
- utils.js: HTTP helpers, text cleaning, debug utilities
- MediaHandler.js: voice/audio + photo processing
- ResponseRenderer.js: dot animation + chunked delivery
- MessageProcessor.js: PTY + ConversationService routing

TelegramChannel.js reduced to 1046 LOC (-34%).

Polling now groups updates by chatId and processes chats
in parallel (Promise.allSettled), serial within same chat.
