+++ server/channels/telegram/utils.js
@@ -1,0 +1,142 @@
+'use strict';
+
+const https = require('https');
+
+const TELEGRAM_HOST = 'api.telegram.org';
+
+// ── Debug condicional (activar con DEBUG_TELEGRAM=1) ─────────────────────────
+function _tgDebug() { return process.env.DEBUG_TELEGRAM === '1'; }
+function tdbg(scope, ...args) {
+  if (!_tgDebug()) return;
+  console.log(`[TG:DBG:${scope}]`, ...args);
+}
+
+// ── Utilidades HTTP ──────────────────────────────────────────────────────────
+
+function httpsPost(urlPath, body, timeoutMs = 35000) {
+  return new Promise((resolve, reject) => {
+    const data = JSON.stringify(body);
+    const options = {
+      hostname: TELEGRAM_HOST,
+      path: urlPath,
+      method: 'POST',
+      family: 4,
+      headers: {
+        'Content-Type': 'application/json',
+        'Content-Length': Buffer.byteLength(data),
+      },
+    };
+    const req = https.request(options, (res) => {
+      let raw = '';
+      res.on('data', chunk => { raw += chunk; });
+      res.on('end', () => {
+        try { resolve(JSON.parse(raw)); }
+        catch (e) { reject(new Error('Respuesta no es JSON: ' + raw.slice(0, 200))); }
+      });
+    });
+    req.setTimeout(timeoutMs, () => { req.destroy(new Error('Timeout')); });
+    req.on('error', reject);
+    req.write(data);
+    req.end();
+  });
+}
+
+function httpsPostMultipart(urlPath, fields, file, timeoutMs = 35000) {
+  return new Promise((resolve, reject) => {
+    const boundary = '----FormBoundary' + Date.now().toString(36) + Math.random().toString(36).slice(2);
+    const parts = [];
+
+    // Campos de texto
+    for (const [key, value] of Object.entries(fields)) {
+      parts.push(
+        Buffer.from(`--${boundary}\r\nContent-Disposition: form-data; name="${key}"\r\n\r\n${value}\r\n`)
+      );
+    }
+
+    // Archivo binario
+    if (file) {
+      parts.push(
+        Buffer.from(
+          `--${boundary}\r\nContent-Disposition: form-data; name="${file.fieldName}"; filename="${file.filename}"\r\n` +
+          `Content-Type: ${file.contentType}\r\n\r\n`
+        )
+      );
+      parts.push(file.buffer);
+      parts.push(Buffer.from('\r\n'));
+    }
+
+    parts.push(Buffer.from(`--${boundary}--\r\n`));
+    const body = Buffer.concat(parts);
+
+    const options = {
+      hostname: TELEGRAM_HOST,
+      path: urlPath,
+      method: 'POST',
+      family: 4,
+      headers: {
+        'Content-Type': `multipart/form-data; boundary=${boundary}`,
+        'Content-Length': body.length,
+      },
+    };
+    const req = https.request(options, (res) => {
+      let raw = '';
+      res.on('data', chunk => { raw += chunk; });
+      res.on('end', () => {
+        try { resolve(JSON.parse(raw)); }
+        catch (e) { reject(new Error('Respuesta no es JSON: ' + raw.slice(0, 200))); }
+      });
+    });
+    req.setTimeout(timeoutMs, () => { req.destroy(new Error('Timeout')); });
+    req.on('error', reject);
+    req.write(body);
+    req.end();
+  });
+}
+
+function cleanPtyOutput(raw) {
+  let s = raw.replace(/\x1B\[(\d*)C/g, (_, n) => ' '.repeat(Number(n) || 1));
+  s = s
+    .replace(/\x1B\[[0-9;?]*[A-Za-z@]/g, '')
+    .replace(/\x1B\][^\x07\x1B]*(?:\x07|\x1B\\)/g, '')
+    .replace(/\x1B[A-Z\\]/g, '')
+    .replace(/[\x00-\x08\x0E-\x1F\x7F]/g, '');
+  const lines = s.split('\n').map(line => {
+    const segs = line.split('\r');
+    let rendered = segs[0] || '';
+    for (let i = 1; i < segs.length; i++) {
+      const seg = segs[i];
+      if (seg.length >= rendered.length) rendered = seg;
+      else if (seg.length > 0) rendered = seg + rendered.slice(seg.length);
+    }
+    return rendered.trimEnd();
+  });
+  const filtered = lines.filter(line => {
+    const t = line.trim();
+    if (!t) return false;
+    if (/^[─━═\-─]{4,}$/.test(t)) return false;
+    if (/^[▐▛▜▌▝▘█▙▟▄▀■]+/.test(t)) return false;
+    if (/^\?.*shortcuts/.test(t)) return false;
+    if (/^ctrl\+/.test(t)) return false;
+    if (/^❯\s*$/.test(t)) return false;
+    return true;
+  });
+  return filtered.join('\n').replace(/\n{3,}/g, '\n\n').trim();
+}
+
+function stripAnsi(str) { return cleanPtyOutput(str); }
+
+function chunkText(text, size = 4096) {
+  const chunks = [];
+  for (let i = 0; i < text.length; i += size) chunks.push(text.slice(i, i + size));
+  return chunks.length > 0 ? chunks : [''];
+}
+
+module.exports = {
+  httpsPost,
+  httpsPostMultipart,
+  cleanPtyOutput,
+  stripAnsi,
+  chunkText,
+  tdbg,
+  TELEGRAM_HOST,
+};
